+++ np.Shared/CHANGELOG.md
@@ -1,6 +1,12 @@
-# Changes to np.Shared
+# Changes to 🤝 Shared Resources plugin
 
 See [Shared Plugin's README](https://github.com/NotePlan/plugins/blob/main/np.Shared/README.md) for details on this plugin.
+
+## [0.3.0] - 2023-02-22 (@jgclark)
+
+### Added
+
+- function `checkForWantedResources(fileList?)` for plugins to use. See its jsdoc for more details.
 
 ## [0.2.0] - 2023-02-18 (@dwertheimer)